LayoutTests/js/constant-count-expected.txt

This test checks exceptional cases for constant counting in the parser.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a is undefined
PASS f() is undefined
PASS successfullyParsed is true

TEST COMPLETE
